+++ MasterThesis/RePair_v2/SymbolRecord.h
@@ -2,16 +2,16 @@
 class SymbolRecord
 {
 public:
-	unsigned int symbol;
+	long symbol;
 	long index;
 	SymbolRecord* previous;
 	SymbolRecord* next;
 
 	SymbolRecord();
-	SymbolRecord(unsigned int s);
-	SymbolRecord(unsigned int s, long index);
-	SymbolRecord(unsigned int s, SymbolRecord * p, SymbolRecord * n);
-	SymbolRecord(unsigned int s, long index, SymbolRecord * p, SymbolRecord * n);
+	SymbolRecord(long s);
+	SymbolRecord(long s, long index);
+	SymbolRecord(long s, SymbolRecord * p, SymbolRecord * n);
+	SymbolRecord(long s, long index, SymbolRecord * p, SymbolRecord * n);
 	~SymbolRecord();
 };
