+++ MasterThesis/unittest_RePair_v2/unittest_RePair_v2_Decoder.cpp
@@ -5,11 +5,11 @@ using namespace std;
 
 TEST(decoder, diddyAll)
 {
-	unordered_map<unsigned int, unordered_map<unsigned int, PairTracker>> activePairs;
+	unordered_map<long, unordered_map<long, PairTracker>> activePairs;
 	vector<SymbolRecord*> sequenceArray;
 	vector<PairRecord*> priorityQueue;
-	unordered_map<unsigned int, Pair> dictionary;
-	unsigned int symbols(initialSymbolValue);//256
+	unordered_map<long, Pair> dictionary;
+	long symbols(initialSymbolValue);//256
 
 	Initializer init;
 	Conditions c;
@@ -25,9 +25,9 @@ TEST(decoder, diddyAll)
 	int priorityQueueSize;
 	int blockSize;
 	blockSize = 1048576;
-	unordered_set<unsigned int> terminals;
+	unordered_set<long> terminals;
 	vector<CompactPair> pairs;
-	unordered_map <unsigned int, unordered_map<unsigned int, unsigned int>> indices;
+	unordered_map <long, unordered_map<long, long>> indices;
 	string filename = input1;
 	ifstream file(filename);
 	bool firstBlock = true;
